requirements.txt
Pillow~=10.0.0
setuptools~=52.0.0
waveshare-epaper==1.2.0
python-dateutil==2.8.2